Simple_Chatbot/requirements.txt
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.7
python-dotenv==1.0.1
